+++ mrp_multi_level/readme/HISTORY.rst
@@ -1,3 +1,9 @@
+11.0.1.1.0 (2018-08-30)
+~~~~~~~~~~~~~~~~~~~~~~~
+
+* [FIX] Consider *Qty Multiple* on product to propose the quantity to procure.
+  (`#297 <https://github.com/OCA/manufacture/pull/297>`_)
+
 11.0.1.0.1 (2018-08-03)
 ~~~~~~~~~~~~~~~~~~~~~~~